pyproject.toml
[project]
name = "projectwise-quart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "jsonschema>=4.25.0",
    "mcp[cli]>=1.12.4",
    "mem0ai>=0.1.115",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "quart>=0.20.0",
    "sqlalchemy>=2.0.43",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
]
